Fix race condition
